Added functions to parse instruction lines and execute on the motors

+++ src/plotter.c
@@ -31,7 +31,7 @@ void test_servo(int x) {
     DDRB = 0x00 | (1 << DDB1) | (1 << DDB2);
     TCCR1A |= (1 << COM1A1) | (1 << WGM11) | (1 << WGM12) | (1 << WGM13);
 	TCCR1B |= (1 << CS12);
-    ICR1 = F_CPU / 50;  
+    ICR1 = (unsigned int)F_CPU / 50;  
     OCR1A = ICR1 - x;
 }
